// src/app/Models/CommandRequest.ts
export interface CommandRequest {
    value?: string;
    displayValue?: string;
    dateCreated?: Date;
    displayFrom?: string;
    displayRecipient?: string;
    displayRequester?: string;
    reason?: string;
    recipient?: string;
    requester?: string;
    reviews?: any;
    secondaryValue?: string;
    type?: string;
}